+++ tests/test_console.py
@@ -360,6 +360,28 @@ class TestAquaConsole():
             run_aqua(['-v', 'fixes', 'remove', 'ciccio.yaml'])
             assert excinfo.value.code == 1
 
+        # set the grids path in the config-aqua.yaml
+        run_aqua(['-v', 'grids', 'set', os.path.join(mydir, 'pippo')])
+        assert os.path.exists(os.path.join(mydir, 'pippo', 'grids'))
+        assert os.path.exists(os.path.join(mydir, 'pippo', 'areas'))
+        assert os.path.exists(os.path.join(mydir, 'pippo', 'weights'))
+        config_file = load_yaml(os.path.join(mydir, '.aqua', 'config-aqua.yaml'))
+        assert config_file['paths'] == {
+            'grids': os.path.join(mydir, 'pippo', 'grids'),
+            'areas': os.path.join(mydir, 'pippo', 'areas'),
+            'weights': os.path.join(mydir, 'pippo', 'weights')
+        }
+        
+        # set the grids path in the config-aqua.yaml with the block already existing
+        run_aqua(['-v', 'grids', 'set', os.path.join(mydir, 'pluto')])
+        assert os.path.exists(os.path.join(mydir, 'pluto', 'grids'))
+        config_file = load_yaml(os.path.join(mydir, '.aqua', 'config-aqua.yaml'))
+        assert config_file['paths'] == {
+            'grids': os.path.join(mydir, 'pluto', 'grids'),
+            'areas': os.path.join(mydir, 'pluto', 'areas'),
+            'weights': os.path.join(mydir, 'pluto', 'weights')
+        }
+
         # uninstall everything
         run_aqua_console_with_input(['uninstall'], 'yes')
         assert not os.path.exists(os.path.join(mydir, '.aqua'))